backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
dj-stripe==2.7.3
Django==4.2.2
django-cleanup==8.0.0
django-cors-headers==4.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
drf-stripe-subscription==1.2.0
idna==3.4
inflection==0.5.1
jsonfield==3.1.0
jsonschema==4.17.3
Pillow==9.5.0
pydantic==1.10.9
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
stripe==4.2.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==6.4.0
